Separated Blockchain and BlockchainState in blockchain.rs.
