+++ blockchain-albatross/src/lib.rs
@@ -18,11 +18,70 @@ extern crate nimiq_tree_primitives as tree_primitives;
 extern crate nimiq_utils as utils;
 extern crate nimiq_vrf as vrf;
 
-pub use blockchain::{Blockchain, ForkEvent};
+use block::{Block, BlockError, ForkProof};
+pub use blockchain::Blockchain;
 
 pub mod blockchain;
+pub mod blockchain_state;
 pub mod chain_info;
 pub mod chain_store;
 pub mod reward;
 pub mod slots;
 pub mod transaction_cache;
+
+pub type PushResult = blockchain_base::PushResult;
+pub type PushError = blockchain_base::PushError<BlockError>;
+pub type BlockchainEvent = blockchain_base::BlockchainEvent<Block>;
+
+pub enum OptionalCheck<T> {
+    Some(T),
+    None,
+    Skip,
+}
+
+impl<T> OptionalCheck<T> {
+    pub fn is_some(&self) -> bool {
+        if let OptionalCheck::Some(_) = self {
+            true
+        } else {
+            false
+        }
+    }
+
+    pub fn is_none(&self) -> bool {
+        if let OptionalCheck::None = self {
+            true
+        } else {
+            false
+        }
+    }
+
+    pub fn is_skip(&self) -> bool {
+        if let OptionalCheck::Skip = self {
+            true
+        } else {
+            false
+        }
+    }
+}
+
+impl<T> From<Option<T>> for OptionalCheck<T> {
+    fn from(opt: Option<T>) -> Self {
+        match opt {
+            Some(t) => OptionalCheck::Some(t),
+            None => OptionalCheck::None,
+        }
+    }
+}
+
+#[derive(Debug, Eq, PartialEq)]
+enum ChainOrdering {
+    Extend,
+    Better,
+    Inferior,
+    Unknown,
+}
+
+pub enum ForkEvent {
+    Detected(ForkProof),
+}
